+++ src/utils/call.ts
@@ -1,0 +1,11 @@
+import { Contract } from '@ethersproject/contracts';
+
+export async function call(provider, abi: any[], call: any[], options?) {
+  const contract = new Contract(call[0], abi, provider);
+  try {
+    const params = call[2] || [];
+    return await contract[call[1]](...params, options || {});
+  } catch (e) {
+    return Promise.reject(e);
+  }
+}
